Aliases for rounding funcs + stop (int) parsing
- Adds aliases for floor, ceil and round (fl, cl, and rn)
- Adds underscore as character allowed in variable name
- Removes autoparsing to (int) in evaluate

+++ ExpressionParser/ExpressionInterpreter.php
@@ -11,7 +11,7 @@
                 die($e->getMessage());
             }
 
-            return (int)(new Interpreter($variables))->interpret($expressions);
+            return (new Interpreter($variables))->interpret($expressions);
         }
 
     }
